File structure refactored
