+++ vizro-core/src/vizro/_vizro.py
@@ -1,3 +1,5 @@
+# isort: skip_file
+# required to prevent circular import issues due to sequence of imports
 import logging
 import os
 from pathlib import Path
@@ -9,7 +11,7 @@ from dash import Dash
 from vizro._constants import STATIC_URL_PREFIX
 from vizro.managers import data_manager, model_manager
 from vizro.models import Dashboard
-from vizro.models._action._callback_utils import _make_app_callbacks
+from vizro.actions._action_loop._build_app_callbacks import _build_app_callbacks
 
 logger = logging.getLogger(__name__)
 
@@ -49,7 +51,7 @@ class Vizro:
         self._pre_build()
 
         self.dash.layout = dashboard.build()
-        _make_app_callbacks()
+        _build_app_callbacks()
 
         return self
