+++ vizro-core/src/vizro/models/_action/__init__.py
@@ -1,0 +1,3 @@
+from vizro.models._action._action import Action
+
+__all__ = ["Action"]
