+++ vizro-core/tests/unit/vizro/actions/test_filter_interaction_action.py
@@ -4,10 +4,10 @@ from dash._utils import AttributeDict
 
 import vizro.models as vm
 from vizro.actions import filter_interaction
-from vizro.managers import model_manager
-from vizro.models._action._actions_utils import (
+from vizro.actions._actions_utils import (
     CallbackTriggerDict,
 )
+from vizro.managers import model_manager
 
 
 @pytest.fixture
