+++ vizro-core/tests/unit/vizro/actions/test_on_page_load_action.py
@@ -6,10 +6,10 @@ import vizro.models as vm
 import vizro.plotly.express as px
 from vizro._constants import ON_PAGE_LOAD_ACTION_PREFIX
 from vizro._themes import dark, light
-from vizro.managers import model_manager
-from vizro.models._action._actions_utils import (
+from vizro.actions._actions_utils import (
     CallbackTriggerDict,
 )
+from vizro.managers import model_manager
 
 
 @pytest.fixture
